--- Assets/Scripts/TankOne.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4307d126ac67ffc4da2c7e0c031a53c7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
