+++ Assets/Scripts/greentank.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7967e0442b8672e45942d0729871ac3f
+guid: fbe9d5738820c02479c80468d2c8a572
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
